Added build pipeline and made changes to satisfy scriptanalyzer.

+++ Private/Set-NSTempDirectory.ps1
@@ -1,10 +1,15 @@
 ﻿Function Set-NSTempDirectory{
-    [CmdletBinding()]
-    $TempPath = "$Env:Temp\PowerShellNS"
-    if(-not (Test-Path $TempPath)){
-        New-Item -Path $TempPath -ItemType Directory
-    }else{
-        Get-ChildItem $TempPath | Remove-Item -Force -Recurse
+    [cmdletbinding(SupportsShouldProcess)]
+    Param(
+        $TempPath = "$Env:Temp\PowerShellNS"
+    )
+    
+    if ($PSCmdlet.ShouldProcess($TempPath)) { 
+        if(-not (Test-Path $TempPath)){
+            New-Item -Path $TempPath -ItemType Directory
+        }else{
+            Get-ChildItem $TempPath | Remove-Item -Force -Recurse
+        }
+        return $TempPath
     }
-    return $TempPath
 }
